gg/requirements.txt
click==6.7
google==2.0.1
boto3==1.9.35
python-dotenv==0.9.1
tabulate==0.8.2
pymongo==3.7.2
requests==2.20.0
